feat: Make gpt4all work offline.
Gpt4all is an offline model and can work
without internet. Right now allow_download
flags download list of model, due to which
its not working offline.
This PR removes the flag for now, but in
future we will add support in the yaml config
so that users can configure themselves.