Add AWS DynamoDB implementation for conversation history and graph database
This commit adds support for AWS DynamoDB as a supported database for conversation and graph history. It adds tests and a sample of the support along with scripts to create the required tables.

In preparing the pull requests, testing found that there were some other changes required. These have also been added here.

Fix test failures due to API changes in MongoDB and Gemini
Add support for neo4j provider in GraphStoreFactory for testing